[refactor] 결제정보표시 컴포넌트 분리
